Intake run always (Requested by Alex)
This change was made so that when the intake is inactive the bottom wheels run inwards to keep the hatch pannel in. I overloaded the "setSpeed" command so that there now is a version that takes in two doubles: "topSpeed" and "bottomSpeed". I then modified the "end" method in the IntakeIn and IntakeOut commands so that they now set the bottom intake wheels to run at half speed when they are inactive.